moved listeners for sheet into Sheet.svelte
